pyproject.toml
[tool.poetry]
name = "LBG Agents"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.12.0"
langchain = "*"
langchain_community = "*"
langchain-google-vertexai = "*"
langchain-google-community = "*"
langchain-text-splitters = "*"
litellm = "^1.66.1"
google-cloud-storage = "^2.18.0"
google-cloud-aiplatform = "*"
google-cloud-pipeline-components = "2.3.0"
google-cloud-dlp = "3.16.0"
google-cloud-logging = "3.10.0"
pydantic-settings = "2.5.2"
Jinja2 = "3.1.4"
kfp = "2.8.0"
fastapi = "0.115"
uvicorn = {extras = ["standard"], version = "0.34.0"}
pyarrow = "16.0.0"
unstructured = {extras = ["doc", "docx", "csv", "md", "epub", "odt", "rst", "rtf", "tsv", "ppt", "pptx", "xlsx", "gcs", "msg", "org"], version = "0.13.7"}
pyxlsb = "1.0.10"
db-dtypes = "1.2.0"
#pypdf = "^4.1.0"
#configparser = "^6.0.1"
#google-cloud-storage-transfer = "^1.11.3"
#google-cloud-functions = "^1.16.3"
pytest = "^8.3.2"
mock = "^5.1.0"
pytest-asyncio = "^0.23.8"
grpcio = "1.64.1"
streamlit = "^1.42"
google-adk = "1.2.1"
pypandoc = "*"
docx2txt = "*"
fpdf = "*"
htmldocx = "*"
streamlit-mermaid = "*"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
ipykernel = "^6.29.3"
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
